Fix cli scripts
